requirements.txt
PyMuPDF
openai
tiktoken
faiss-cpu
python-dotenv
langchain==0.1.14
langchain-community==0.0.30
langchain-openai
